Expose CURLPROTO_* values in protocols table
Add a table "protocols" in the module with protocol names as keys and
CURLPROTO_* defines from curl.h as values. Those values are meant to be
used with the bit32 module in lua 5.2 and with Lua BitOp in lua 5.1 and
luajit.

Exposing protocol values is the first step toward adding support for
CURLOPT_PROTOCOLS and CURLOPT_REDIR_PROTOCOLS from curl_easy_setopt.

Signed-off-by: Davide "FunkyAss" Del Zompo <davide.delzompo@gmail.com>